# perf: cache product bundle items at document level