**O(N) => O(1)**

This should help in faster submissions and validations.

### Before

```py
In [1]: doc = frappe.get_doc('Sales Invoice', 'SINV-24-00007')

In [2]: len(doc.items)
Out[2]: 16

In [3]: %timeit -n 10 doc.get_item_list()
2.79 ms ± 1.31 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [4]: %timeit -n 10 doc.get_item_list()
2.48 ms ± 629 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

### After

```py
In [1]: doc = frappe.get_doc('Sales Invoice', 'SINV-24-00007')

In [2]: len(doc.items)
Out[2]: 16

In [3]: %timeit -n 10 doc.get_item_list()
The slowest run took 11.67 times longer than the fastest. This could mean that an intermediate result is being cached.
162 µs ± 210 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [4]: %timeit -n 10 doc.get_item_list()
66.1 µs ± 34.4 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```